pyproject.toml
[project]
name = "universal-storefront"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pg8000 (>=1.31.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]